+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ﻿# Changelog
 All notable changes to this project will be documented in this file.
 ---
+### `12.0.0.5` (2026-04-08)
+**Fixes**
+- Fixed issue with QuestInfo for targets inside instances (secret value).
+- Fixed issue that caused the start up notification to be popped for release notes that are not that interested.
+
+---
 ### `12.0.0.4` (2026-03-25)
 **Fixes**
 - Fixed several issues caused by Blizzards recent API changes to prevents certain information to be read or manipulated during combat. In this case the AFK, DND and PVP status of players, and tooltip information.